Fix #12685: Add tests for maven-war-plugin and maven-ear-plugin upgrades, extract Quarkus tests
* Fix #12685: Add maven-war-plugin and maven-ear-plugin to plugin upgrade list

Both plugins use reflection on java.util.Properties internals (via
XStream/plexus-archiver) in older versions, which is blocked by the
JDK 17+ module system. Since Maven 4 requires JDK 17+, mvnup should
upgrade these plugins to compatible versions (>= 3.4.0).


* fix: trim test file to stay within 2000-line Checkstyle limit


* refactor: extract Quarkus plugin tests to separate file

Move the Quarkus-specific tests from PluginUpgradeStrategyTest to a new
PluginUpgradeQuarkusTest class to stay well within the 2000-line
Checkstyle limit and leave room for future plugin additions.


---------
